+++ src/hooks/use-map.tsx
@@ -1,6 +1,6 @@
 import {useEffect, useState, MutableRefObject, useRef} from 'react';
 import {Map, TileLayer} from 'leaflet';
-import { AmsterdamCenterCoords } from '../types/offer';
+import { AmsterdamCenterCoords } from '../types/map';
 import { TILE_LAYER, COPYRIGHT } from '../const';
 
 export default function useMap (
@@ -11,26 +11,29 @@ export default function useMap (
   const isRenderedRef = useRef<boolean>(false);
 
   useEffect(() => {
-    if(mapRef.current !== null && !isRenderedRef.current){
-      const instance = new Map(mapRef.current, {
-        center: {
-          lat: location.lat,
-          lng: location.lng
-        },
-        zoom: location.zoom
-      });
-      const layer = new TileLayer(
-        TILE_LAYER,
-        {
-          attribution:
-          COPYRIGHT
-        }
-      );
-      instance.addLayer(layer);
-
-      setMap(instance);
-      isRenderedRef.current = true;
+    if(mapRef.current === null || isRenderedRef.current){
+      return;
     }
+
+    const instance = new Map(mapRef.current, {
+      center: {
+        lat: location.lat,
+        lng: location.lng
+      },
+      zoom: location.zoom
+    });
+    const layer = new TileLayer(
+      TILE_LAYER,
+      {
+        attribution:
+          COPYRIGHT
+      }
+    );
+    instance.addLayer(layer);
+
+    setMap(instance);
+    isRenderedRef.current = true;
+
   }, [mapRef, location]);
 
   return map;
