+++ src/types/map.ts
@@ -1,0 +1,15 @@
+import { Offer, Offers } from './offer';
+
+export type AmsterdamCenterCoords = {
+  title: string;
+  lat: number;
+  lng: number;
+  zoom: number;
+}
+
+export type MapProps ={
+  block: string;
+  location: AmsterdamCenterCoords;
+  offers: Offers;
+  selectedOffer: Offer | undefined;
+}
